requirements.txt
opencv-python
streamlit
PyMuPDF
numpy
Pillow
